requirements.txt
playwright
google-genai
python-dotenv
pydantic
rich
pytest
requests
jinja2
langgraph